smashed the supply and borrow desktop table files
