+++ src/screens/Dashboard/AccountSupplyAndBorrow/DesktopAccountSupplyAndBorrow/AccountBorrowTable.tsx
@@ -14,7 +14,7 @@ import { useNavigate } from "react-router-dom";
 import { ROUTES } from "@src/constants";
 import { useTheme } from "@emotion/react";
 import { TPoolStats } from "@src/stores/LendStore";
-import { useSupplyAndBorrowVM } from "./SupplyAndBorrowVM";
+import { useSupplyAndBorrowVM } from "@screens/Dashboard/AccountSupplyAndBorrow/DesktopAccountSupplyAndBorrow/SupplyAndBorrowVM";
 
 interface IProps {}
 
@@ -38,126 +38,12 @@ const TableWrap = styled.div<{ sort?: boolean }>`
   }
 `;
 
-const DesktopAccountSupplyAndBorrow: React.FC<IProps> = () => {
+const AccountBorrowTable: React.FC<IProps> = () => {
   const { lendStore } = useStores();
-  const [filteredSupplies, setFilteredSupplies] = useState<any[]>([]);
   const [filteredBorrows, setFilteredBorrows] = useState<any[]>([]);
   const theme = useTheme();
   const vm = useSupplyAndBorrowVM();
   const navigate = useNavigate();
-
-  const supplyColumns = useMemo(
-    () => [
-      { Header: "Asset", accessor: "asset" },
-      { Header: "", accessor: "emptyCell" },
-      {
-        Header: () => (
-          <Row
-            style={{ cursor: "pointer" }}
-            onClick={() =>
-              lendStore.accountSupply.length > 1
-                ? vm.selectSort("selfSupply", true)
-                : null
-            }
-            justifyContent="flex-end"
-          >
-            <Text size="medium" fitContent nowrap>
-              Supplied
-            </Text>
-            {lendStore.accountSupply.length > 1 ? (
-              <img
-                src={theme.images.icons.group}
-                alt="group"
-                className={
-                  vm.sortSupply === "selfSupply"
-                    ? "sort-icon-active"
-                    : "sort-icon"
-                }
-              />
-            ) : (
-              <SizedBox width={24} />
-            )}
-          </Row>
-        ),
-        accessor: "supplied"
-      },
-      {
-        Header: () => (
-          <Row
-            style={{ cursor: "pointer" }}
-            onClick={() =>
-              lendStore.accountSupply.length > 1
-                ? vm.selectSort("supplyAPY", true)
-                : null
-            }
-            justifyContent="flex-end"
-          >
-            <Tooltip
-              content={
-                <Text textAlign="left">
-                  Annual interest paid to investors taking into account
-                  compounding.
-                </Text>
-              }
-            >
-              <Text
-                style={{ textDecoration: "underline dotted" }}
-                size="medium"
-                fitContent
-                nowrap
-              >
-                Supply APY
-              </Text>
-            </Tooltip>
-            {lendStore.accountSupply.length > 1 ? (
-              <img
-                src={theme.images.icons.group}
-                alt="group"
-                className={
-                  vm.sortSupply === "supplyAPY"
-                    ? "sort-icon-active"
-                    : "sort-icon"
-                }
-              />
-            ) : (
-              <SizedBox width={24} />
-            )}
-          </Row>
-        ),
-        accessor: "supplyApy"
-      },
-      {
-        Header: () => (
-          <Row
-            style={{ cursor: "pointer" }}
-            onClick={() => vm.selectSort("dailyIncome", true)}
-            justifyContent="flex-end"
-          >
-            <Text size="medium" fitContent nowrap>
-              Daily income
-            </Text>
-            {lendStore.accountSupply.length > 1 ? (
-              <img
-                src={theme.images.icons.group}
-                alt="group"
-                className={
-                  vm.sortSupply === "dailyIncome"
-                    ? "sort-icon-active"
-                    : "sort-icon"
-                }
-              />
-            ) : (
-              <SizedBox width={24} />
-            )}
-          </Row>
-        ),
-        accessor: "dailyIncome"
-      },
-      { Header: "", accessor: "supplyBtn" },
-      { Header: "", accessor: "withdrawBtn" }
-    ],
-    [vm, theme.images.icons.group, lendStore.accountSupply.length]
-  );
 
   const openModal = useCallback(
     (
@@ -172,116 +58,6 @@ const DesktopAccountSupplyAndBorrow: React.FC<IProps> = () => {
     [navigate]
   );
 
-  useMemo(() => {
-    const data = vm
-      .sortData(lendStore.accountSupply, true)
-      .map((s: TPoolStats) => ({
-        onClick: () =>
-          navigate(
-            ROUTES.DASHBOARD_TOKEN_DETAILS.replace(
-              ":poolId",
-              lendStore.pool.address
-            ).replace(":assetId", s.assetId)
-          ),
-        asset: (
-          <Row alignItems="center">
-            <SquareTokenIcon size="small" src={s.logo} alt="logo" />
-            <SizedBox width={16} />
-            <Column>
-              <Text size="small" fitContent>
-                {s.symbol}
-              </Text>
-              <Text type="secondary" size="small" fitContent>
-                $ {s.prices.max.toBigFormat(2)}
-              </Text>
-            </Column>
-          </Row>
-        ),
-        emptyCell: <SizedBox width={140} />,
-        supplied: (
-          <Column crossAxisSize="max">
-            <Text weight={500} textAlign="right" size="medium">
-              {BN.formatUnits(s.selfSupply, s.decimals).toBigFormat(2) +
-                ` ${s.symbol}`}
-            </Text>
-            <Text textAlign="right" size="small" type="secondary">
-              ${" "}
-              {BN.formatUnits(s.selfSupply, s.decimals)
-                .times(s.prices.min)
-                .toBigFormat(2)}
-            </Text>
-          </Column>
-        ),
-        supplyApy: s.supplyAPY.toBigFormat(2) + "%",
-        dailyIncome: (
-          <Column crossAxisSize="max">
-            <Text weight={500} textAlign="right" size="medium">
-              {BN.formatUnits(s.dailyIncome, s.decimals).toBigFormat(6) +
-                ` ${s.symbol}`}
-            </Text>
-            <Text textAlign="right" size="small" type="secondary">
-              ${" "}
-              {BN.formatUnits(s.dailyIncome, s.decimals)
-                .times(s.prices.min)
-                .toBigFormat(6)}
-            </Text>
-          </Column>
-        ),
-        supplyBtn: vm.isSupplyDisabled(s) ? (
-          <Tooltip
-            content={
-              <Text textAlign="left">Maximum total supply is reached</Text>
-            }
-          >
-            <Button
-              kind="secondary"
-              size="medium"
-              fixed
-              disabled={true}
-              onClick={(e) =>
-                openModal(e, lendStore.poolId, "supply", s.assetId)
-              }
-              style={{ width: "100px", margin: "0 auto" }}
-            >
-              Supply
-            </Button>
-          </Tooltip>
-        ) : (
-          <Button
-            kind="secondary"
-            size="medium"
-            fixed
-            onClick={(e) => openModal(e, lendStore.poolId, "supply", s.assetId)}
-            style={{ width: "100px", margin: "0 auto" }}
-          >
-            Supply
-          </Button>
-        ),
-        withdrawBtn: (
-          <Button
-            kind="secondary"
-            size="medium"
-            fixed
-            onClick={(e) =>
-              openModal(e, lendStore.poolId, "withdraw", s.assetId)
-            }
-            style={{ width: "100px", margin: "0 auto" }}
-          >
-            Withdraw
-          </Button>
-        )
-      }));
-    setFilteredSupplies(data);
-  }, [
-    vm.sortModeSupply,
-    vm.sortSupply,
-    lendStore.pool.address,
-    lendStore.accountSupply,
-    lendStore.poolId,
-    openModal,
-    navigate
-  ]);
-
   //-------------
   const borrowColumns = useMemo(
     () => [
@@ -291,11 +67,10 @@ const DesktopAccountSupplyAndBorrow: React.FC<IProps> = () => {
         Header: () => (
           <Row
             style={{ cursor: "pointer" }}
-            onClick={() =>
-              lendStore.accountBorrow.length > 1
-                ? vm.selectSort("selfBorrow", false)
-                : null
-            }
+            onClick={() => {
+              console.log("To be repaid");
+              vm.selectSort("selfBorrow", false);
+            }}
             justifyContent="flex-end"
           >
             <Text size="medium" fitContent nowrap>
@@ -480,30 +255,17 @@ const DesktopAccountSupplyAndBorrow: React.FC<IProps> = () => {
       }));
     setFilteredBorrows(data);
   }, [
-    vm.sortModeBorrow,
     vm.sortBorrow,
+    vm.sortModeBorrow,
     lendStore.pool.address,
     lendStore.poolId,
     lendStore.accountBorrow,
     openModal,
     navigate
   ]);
+  //todo fix dependency problem
   return (
     <Root>
-      {lendStore.accountSupply.length > 0 && (
-        <TableWrap sort={vm.sortModeSupply === "descending"}>
-          <Text weight={500} type="secondary">
-            My supply
-          </Text>
-          <SizedBox height={8} />
-          <Table
-            style={{ width: "100%" }}
-            columns={supplyColumns}
-            data={filteredSupplies}
-          />
-          <SizedBox height={24} />
-        </TableWrap>
-      )}
       {lendStore.accountBorrow.length > 0 && (
         <TableWrap sort={vm.sortModeBorrow === "descending"}>
           <Text weight={500} type="secondary">
@@ -521,4 +283,4 @@ const DesktopAccountSupplyAndBorrow: React.FC<IProps> = () => {
     </Root>
   );
 };
-export default observer(DesktopAccountSupplyAndBorrow);
+export default observer(AccountBorrowTable);
