+++ src/screens/Dashboard/AccountSupplyAndBorrow/DesktopAccountSupplyAndBorrow/index.tsx
@@ -1,0 +1,3 @@
+import DesktopAccountSupplyAndBorrow from "./DesktopAccountSupplyAndBorrow";
+
+export default DesktopAccountSupplyAndBorrow;
